requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.14.0
mysqlclient==2.2.4
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
gunicorn==20.1.0
psycopg2==2.8.6